Complete CoinData separate component
